\name{exampleDat}
\alias{exampleDat}
\docType{data}
\title{example longitudinal ordinal data}
 \description{
  The data contains 10 variables, 20 samples, for 10 time points.
  }
\usage{
data(exampleDat)
}
\format{
   Data is in longitudinal format with 10 columns, 200 rows (see R package longitudinal).
}

\examples{
# load "longitudinal" library
library(longitudinal)

# load data sets
data(exampleDat)
}