\name{tsnetwork-package}
\alias{tsnetwork-package}
\docType{package}
\title{
 Learning instantaneous networks and dynamic networks for mixed discrete-continuous time-series data. 
}

\description{
The \code{R} package \pkg{tsnetwork} reconstructs instantaneous (undirected) and dynamic (directed) networks from 
 repeated multivariate mixed discrete-continuous or ordinal time series data. 
 This function computes sparse autoregressive coefficient and precision matrices for time
 series chain graphical models.
}

\author{
Maintainers: Pariya Behrouzi \email{pariya.behrouzi@gmail.com}

}

\references{
\cite{Pariya Behrouzi, Fentaw Abegaz and Ernst Wit (2018). Dynamic Chain Graph Models for Ordinal Time Series Data. 
\emph{Arxiv}. \bold{14}, 3: 586-599.}
}


\keyword{package}